ADICIONADO CARGA DE FT REPROVACOES E SCRIPTS DO BANCO

+++ Etl_Avaliacao1/Models/Operacional/Reprovacao.cs
@@ -13,7 +13,7 @@ namespace Etl_Avaliacao1.Models.Operacional
         public int IdDisciplina { get; set; }
         public int IdCurso { get; set; }
         public bool Cotista { get; set; }
-        public long CargaHoraria { get; set; }
+        public int CargaHoraria { get; set; }
         public int Nota { get; set; }
         public int Faltas { get; set; }
     }
